fix(event-processor): Fix SQLx type mismatch in polling fallback
The UnprocessedEvent struct had chain_id as i64, but the PostgreSQL
view unprocessed_events defines chain_id as INTEGER (i32). This
caused SQLx to fail with "Failed to query unprocessed events" error.

Fixed by changing chain_id type from i64 to i32 to match the database
schema.

+++ rust-backend/crates/event-processor/src/polling_fallback.rs
@@ -52,7 +52,7 @@ const MAX_POLLING_ITERATIONS: u64 = 1_000_000;
 #[derive(Debug, Clone, FromRow)]
 struct UnprocessedEvent {
     id: String,
-    chain_id: i64,
+    chain_id: i32, // matches INTEGER in PostgreSQL view
     block_number: i64,
     #[allow(dead_code)] // Used in SQL query and debug logs
     registry: String,
